fix: move all http interceptor related stuff in its own file
