+++ mod.ts
@@ -4,6 +4,7 @@ export * from "./src/cache.ts";
 export * from "./src/dialog.ts";
 export * from "./src/element_handle.ts";
 export * from "./src/file_chooser.ts";
+export * from "./src/interceptor.ts";
 export * from "./src/locator.ts";
 export * from "./src/keyboard/mod.ts";
 export * from "./src/mouse.ts";
